+++ R/std_coeff_fun.R
@@ -54,85 +54,9 @@ sdW <- function(...) {
 }
 
 
-#' @title Get Model Term Names
-#' @description Extract term names from a fitted model object.
-#' @param mod A fitted model object of class \code{"lm"}, \code{"glm"}, or
-#'   \code{"merMod"}, or a list or nested list of such objects.
-#' @param intercept Logical, whether the intercept term should be included.
-#' @param aliased Logical, whether names of aliased terms should be included
-#'   (see Details).
-#' @param list Logical, whether names should be returned as a list, with all
-#'   multi-coefficient terms grouped under their term names.
-#' @param ... Not currently used.
-#' @details Extract term names from a fitted model. Names of terms for which
-#'   coefficients cannot be estimated are also included if \code{aliased = TRUE}
-#'   (default). These may be terms which are perfectly correlated with other
-#'   terms in the model, so that the model design matrix is rank deficient.
-#' @return A character vector or list/nested list of term names.
-#' @examples
-#' ## Term names from Shipley SEM
-#' m <- Shipley.SEM
-#' xNam(m)
-#' xNam(m, intercept = FALSE)  # only predictors
-#'
-#' ## Model with different types of predictor (some multi-coefficient terms)
-#' x1 <- poly(rnorm(100), 2)  # polynomial
-#' x2 <- as.factor(rep(c("a", "b", "c", "d"), each = 25))  # categorical
-#' x3 <- rep(1, 100)  # no variation
-#' m <- lm(rnorm(100) ~ x1 + x2 + x3)
-#' xNam(m)
-#' xNam(m, aliased = FALSE)  # drop term that cannot be estimated (x3)
-#' xNam(m, aliased = FALSE, list = TRUE)  # as named list
-#' @export
-xNam <- function(mod, intercept = TRUE, aliased = TRUE, list = FALSE, ...) {
-
-  m <- mod
-
-  ## Function
-  xNam <- function(m) {
-
-    ## All names as list
-    xn <- labels(terms(m))
-    xn2 <- rownames(summary(m)$coef)
-    xn <- c(xn2[isInt(xn2)], xn)
-    mf <- model.frame(m)
-    XN <- sapply(xn, function(i) {
-      if (i %in% names(mf)) {
-        xi <- mf[, i]
-        xic <- class(xi)
-        if ("factor" %in% xic) {
-          paste0(i, levels(xi)[-1])
-        } else {
-          if ("matrix" %in% xic) {
-            paste0(i, colnames(xi))
-          } else i
-        }
-      } else i
-    }, simplify = FALSE)
-
-    ## Drop intercept?
-    if (!intercept) XN <- XN[!isInt(names(XN))]
-
-    ## Drop aliased terms?
-    if (!aliased) {
-      XN <- XN[sapply(XN, function(i) all(i %in% xn2))]
-    }
-
-    ## Return as list?
-    if (!list) unlist(unname(XN)) else XN
-
-  }
-
-  ## Apply recursively
-  rMapply(xNam, m, SIMPLIFY = FALSE)
-
-}
-
-
 #' @title Get Model Data
 #' @description Extract the data used to fit a model.
-#' @param mod A fitted model object of class \code{"lm"}, \code{"glm"}, or
-#'   \code{"merMod"}, or a list or nested list of such objects.
+#' @param mod A fitted model object, or a list or nested list of such objects.
 #' @param subset Logical. If \code{TRUE}, only observations used to fit the
 #'   model(s) are returned (i.e. missing observations (\code{NA}) are removed).
 #' @param merge Logical. If \code{TRUE}, and \code{mod} is a list or nested
@@ -169,14 +93,15 @@ getData <- function(mod, subset = FALSE, merge = FALSE, ...) {
 
     ## Data from 'data' argument of model call
     mc <- getCall(m)
-    d <- data.frame(eval(mc$data, ...))
+    d <- eval(mc$data, ...)
+    if (!is.null(d)) d <- data.frame(d) else
+      stop("'data' argument of model call not specified.")
 
-    ## All var names from formula
-    f <- c(formula(m), mc$weights, mc$offset)
+    ## All var names from model call
+    f <- c(formula(m), mc$subset, mc$weights, mc$offset, mc$correlation)
     vn <- unlist(lapply(f, all.vars))
-
     if (!all(vn %in% names(d)))
-      stop("'data' must contain all variables used to fit model.")
+      stop("'data' does not contain all model variables.")
 
     ## Subset for model observations?
     if (subset) {
@@ -203,13 +128,110 @@ getData <- function(mod, subset = FALSE, merge = FALSE, ...) {
 }
 
 
+#' @title Get Model Term Names
+#' @description Extract term names from a fitted model object.
+#' @param mod A fitted model object, or a list or nested list of such objects.
+#' @param data An optional dataset used to construct the model frame.
+#' @param intercept Logical, whether the intercept should be included.
+#' @param aliased Logical, whether names of aliased terms should be included
+#'   (see Details).
+#' @param list Logical, whether names should be returned as a list, with all
+#'   multi-coefficient terms grouped under their main term names.
+#' @param ... Arguments to \code{eval} (for evaluating model data).
+#' @details Extract term names from a fitted model. Names of terms for which
+#'   coefficients cannot be estimated are also included if \code{aliased = TRUE}
+#'   (default). These may be terms which are perfectly correlated with other
+#'   terms in the model, so that the model design matrix is rank deficient.
+#' @return A character vector or list/nested list of term names.
+#' @examples
+#' ## Term names from Shipley SEM
+#' m <- Shipley.SEM
+#' xNam(m)
+#' xNam(m, intercept = FALSE)
+#'
+#' ## Model with different types of predictor (some multi-coefficient terms)
+#' d <- data.frame(
+#'   y = rnorm(100),
+#'   x1 = poly(rnorm(100), 2),  # polynomial
+#'   x2 = as.factor(rep(c("a", "b", "c", "d"), each = 25)),  # categorical
+#'   x3 = rep(1, 100)  # no variation
+#' )
+#' m <- lm(y ~ x1.1 + x1.2 + x2 + x3, data = d)
+#' xNam(m)
+#' xNam(m, aliased = FALSE)  # drop term that cannot be estimated (x3)
+#' xNam(m, aliased = FALSE, list = TRUE)  # as named list
+#' @export
+xNam <- function(mod, data = NULL, intercept = TRUE, aliased = TRUE,
+                 list = FALSE, ...) {
+
+  m <- mod; d <- data
+
+  ## Function
+  xNam <- function(m) {
+
+    ## Term names (as vector)
+    tt <- terms(m)
+    xn <- labels(tt)
+    int <- attr(tt, "intercept")
+    if (int) xn <- c("(Intercept)", xn)
+
+    ## Variables
+    if (is.null(d)) d <- getData(m, ...)
+    mf <- model.frame(m, data = d)
+    x <- sapply(mf[names(mf) %in% xn], "[", simplify = FALSE)
+
+    ## Expand factor/matrix terms (list)
+    XN <- sapply(xn, function(i) {
+      if (i %in% names(x)) {
+        j <- c(levels(x[[i]])[-1], colnames(x[[i]]))
+        paste0(i, j)
+      } else i
+    }, simplify = FALSE)
+
+    ## Expand interaction terms involving factors/matrices
+    XN <- sapply(xn, function(i) {
+      if (isInx(i)) {
+        j <- unlist(strsplit(i, ":"))
+        j <- expand.grid(XN[j])
+        apply(j, 1, paste, collapse = ":")
+      } else XN[[i]]
+    }, simplify = FALSE)
+
+    ## If no intercept, add reference level to first factor
+    fac <- sapply(x, function(i) "factor" %in% class(i))
+    if (!int && any(fac)) {
+      f1 <- names(fac[fac][1])
+      XN[[f1]] <- paste0(f1, levels(x[[f1]]))
+    }
+
+    ## Drop intercept?
+    if (!intercept && int) XN <- XN[-1]
+
+    ## Drop aliased terms?
+    if (!aliased) {
+      b <- summary(m)$coef
+      b <- as.matrix(if (isList(b)) b[[1]] else b)
+      XN <- lapply(XN, function(i) i[i %in% rownames(b)])
+      XN <- XN[sapply(XN, length) > 0]
+    }
+
+    ## Return as list?
+    if (!list) unlist(unname(XN)) else XN
+
+  }
+
+  ## Apply recursively
+  rMapply(xNam, m, SIMPLIFY = FALSE)
+
+}
+
+
 #' @title Get Model Response Variable
 #' @description Extract the response variable from a fitted model in the
 #'   original or link scale (for GLM's).
-#' @param mod A fitted model object of class \code{"lm"}, \code{"glm"}, or
-#'   \code{"merMod"}. Alternatively, a numeric vector, corresponding to a
-#'   variable to be transformed. Can also be a list or nested list of such
-#'   objects.
+#' @param mod A fitted model object. Alternatively, a numeric vector,
+#'   corresponding to a variable to be transformed. Can also be a list or nested
+#'   list of such objects.
 #' @param family Optional, the error distribution family containing the link
 #'   function which will be used to transform the response (see
 #'   \code{\link[stats]{family}} for specification details).
@@ -369,6 +391,8 @@ getY <- function(mod, family = NULL, data = NULL, link = FALSE, ...) {
       y <- as.numeric(m)
       setNames(y, names(m))
     }
+    a <- names(attributes(y))
+    attributes(y)[a != "names"] <- NULL
 
     ## Return in original or link scale
     if (isGlm(m) && link || !mod && !is.null(f)) {
@@ -378,7 +402,9 @@ getY <- function(mod, family = NULL, data = NULL, link = FALSE, ...) {
         f <- get(f, mode = "function", envir = parent.frame())
       }
       if (is.function(f)) f <- f()
-      if (is.null(f)) f <- family(m)
+      if (is.null(f)) {
+        f <- if (isBet(m)) m$link$mean else family(m)
+      }
 
       ## Transform response to link scale
       yl <- f$linkfun(y)
@@ -416,10 +442,9 @@ getY <- function(mod, family = NULL, data = NULL, link = FALSE, ...) {
 #' @title Generalised Variance Inflation Factors
 #' @description Calculate generalised variance inflation factors for terms in a
 #'   fitted model via the variance-covariance matrix of coefficients.
-#' @param mod A fitted model object of class \code{"lm"}, \code{"glm"}, or
-#'   \code{"merMod"}, or a list or nested list of such objects.
+#' @param mod A fitted model object, or a list or nested list of such objects.
 #' @param data An optional dataset used to first re-fit the model(s).
-#' @param ... Not currently used.
+#' @param ... Arguments to \code{eval} (for evaluating model data).
 #' @details \code{VIF} calculates generalised variance inflation factors (GVIF)
 #'   as described in Fox & Monette (1992), and also implemented in the
 #'   \code{vif} function in the \pkg{car} package. However, whereas \code{vif}
@@ -447,10 +472,13 @@ getY <- function(mod, family = NULL, data = NULL, link = FALSE, ...) {
 #' VIF(update(m, . ~ . - DD))  # drop DD
 #'
 #' ## Model with different types of predictor (some multi-coefficient terms)
-#' x1 <- poly(rnorm(100), 2)  # polynomial
-#' x2 <- as.factor(rep(c("a", "b", "c", "d"), each = 25))  # categorical
-#' x3 <- rep(1, 100)  # no variation
-#' m <- lm(rnorm(100) ~ x1 + x2 + x3)
+#' d <- data.frame(
+#'   y = rnorm(100),
+#'   x1 = poly(rnorm(100), 2),  # polynomial
+#'   x2 = as.factor(rep(c("a", "b", "c", "d"), each = 25)),  # categorical
+#'   x3 = rep(1, 100)  # no variation
+#' )
+#' m <- lm(y ~ x1.1 + x1.2 + x2 + x3, data = d)
 #' VIF(m)
 #' @export
 VIF <- function(mod, data = NULL, ...) {
@@ -463,32 +491,29 @@ VIF <- function(mod, data = NULL, ...) {
     ## Update model with any supplied data
     if (!is.null(d)) m <- eval(update(m, data = d, evaluate = FALSE))
 
-    ## Term names (all)
-    XN <- xNam(m, intercept = FALSE, list = TRUE)
-    xn <- unlist(unname(XN))
-
-    ## Term names (drop aliased)
-    XN2 <- xNam(m, intercept = FALSE, list = TRUE, aliased = FALSE)
-    xn2 <- unlist(unname(XN2))
+    ## Term names
+    XN <- xNam(m, intercept = FALSE, list = TRUE, ...)
+    xn <- xNam(m, intercept = FALSE, aliased = FALSE, ...)
 
     ## VIF's
-    if (length(xn2) > 1) {
+    if (length(xn) > 1) {
 
       ## T/F for terms as matrices
-      mf <- model.frame(m)
+      if (is.null(d)) d <- getData(m, ...)
+      mf <- model.frame(m, data = d)
       mat <- sapply(names(XN), function(i) {
         if (i %in% names(mf)) class(mf[, i])[1] == "matrix" else FALSE
       })
 
-      ## var-cov & cor matrix
-      V <- as.matrix(vcov(m))[xn2, xn2]
+      ## Var-cov/cor matrix
+      V <- as.matrix(vcov(m))[xn, xn]
       R <- cov2cor(V)
       det.R <- det(R)
 
       ## Function
       VIF <- function(i) {
-        if (all(i %in% xn2)) {
-          ni <- !xn2 %in% i
+        if (all(i %in% xn)) {
+          ni <- !xn %in% i
           Ri <- R[i, i, drop = FALSE]
           Rni <- R[ni, ni, drop = FALSE]
           vif <- det(Ri) * det(Rni) / det.R
@@ -506,7 +531,9 @@ VIF <- function(mod, data = NULL, ...) {
       unlist(unname(vif))
 
     } else {
-      sapply(xn, function(i) if (i %in% xn2) 1 else NA)
+      sapply(unlist(unname(XN)), function(i) {
+        if (i %in% xn) 1 else NA
+      })
     }
 
   }
@@ -522,11 +549,11 @@ VIF <- function(mod, data = NULL, ...) {
 #'   defined as the squared multiple correlation between the observed and fitted
 #'   values for the response variable. 'Adjusted' and 'Predicted' versions are
 #'   also calculated (see Details).
-#' @param mod A fitted model object of class \code{"lm"}, \code{"glm"}, or
-#'   \code{"merMod"}, or a list or nested list of such objects.
+#' @param mod A fitted model object, or a list or nested list of such objects.
 #' @param data An optional dataset used to first re-fit the model(s).
 #' @param adj,pred Logical. If \code{TRUE} (default), adjusted and/or predicted
-#'   R squared are also returned.
+#'   R-squared are also returned (the latter is not available for all model
+#'   types).
 #' @param re.form For mixed models of class \code{"merMod"}, the formula for
 #'   random effects to condition on when generating fitted values used in the
 #'   calculation of R-squared. Defaults to \code{NULL}, meaning all random
@@ -672,7 +699,6 @@ VIF <- function(mod, data = NULL, ...) {
 #' # lm: TRUE; glm: "Mean relative difference: 1.977725e-06"
 #' cor(cvf1, cvf2)
 #' # lm: 1; glm: 0.9999987
-#'
 #' }
 #'
 #' # NOTE: comparison not tested here for mixed models, as hierarchical data can
@@ -695,9 +721,11 @@ R2 <- function(mod, data = NULL, adj = TRUE, pred = TRUE, re.form = NULL,
     if (!is.null(d)) m <- eval(update(m, data = d, evaluate = FALSE))
 
     ## R squared
+    b <- summary(m)$coef
+    b <- as.matrix(if (isList(b)) b[[1]] else b)
     i <- attr(terms(m), "intercept")
-    k <- nrow(summary(m)$coef) - i
-    if (isMerMod(m)) k <- k + length(m@theta)
+    k <- nrow(b) - i
+    if (isMer(m)) k <- k + length(m@theta)
     R2 <- if (k > 0) {
       y <- getY(m)
       n <- length(y)
@@ -715,20 +743,22 @@ R2 <- function(mod, data = NULL, adj = TRUE, pred = TRUE, re.form = NULL,
       if (R2 > 0) {
         R2a <- 1 - ((n - 3) * (1 - R2) / (n - k - i)) *
           (1 + (2 * (1 - R2) / (n - k - 2.3)))
-        # R2a <- 1 - (1 - R2) * (n - i) / (n - k - i)  # standard formula (Ezekiel/Wherry)
+        # R2a <- 1 - (1 - R2) * (n - i) / (n - k - i)  # 'standard' formula
         if (R2a > 0) R2a else 0
       } else 0
     }
 
     ## Predictive R squared
     R2p <- if (pred) {
-      if (R2 > 0) {
-        hii <- suppressWarnings(hatvalues(m))
-        s <- hii < 1
-        f <- y - (y - f) / (1 - hii)
-        Rp <- cov.wt(cbind(y, f)[s, ], w[s], cor = TRUE)$cor[1, 2]
-        if (Rp > 0) Rp^2 else 0
-      } else 0
+      if (!isGls(m)) {
+        if (R2 > 0) {
+          hii <- suppressWarnings(hatvalues(m)[s])
+          s <- hii < 1
+          f <- y - (y - f) / (1 - hii)
+          Rp <- cov.wt(cbind(y, f)[s, ], w[s], cor = TRUE)$cor[1, 2]
+          if (Rp > 0) Rp^2 else 0
+        } else 0
+      } else NA
     }
 
     ## Return values
@@ -859,8 +889,7 @@ avgEst <-  function(est, weights = "equal", est.names = NULL, ...) {
 #' @title Standardised Coefficients
 #' @description Calculate fully standardised model coefficients in standard
 #'   deviation units, adjusted for multicollinearity among predictors.
-#' @param mod A fitted model object of class \code{"lm"}, \code{"glm"}, or
-#'   \code{"merMod"}, or a list or nested list of such objects.
+#' @param mod A fitted model object, or a list or nested list of such objects.
 #' @param weights An optional numeric vector of weights to use for model
 #'   averaging, or a named list of such vectors. The former should be supplied
 #'   when \code{mod} is a list, and the latter when it is a nested list (with
@@ -1020,25 +1049,30 @@ stdCoeff <- function(mod, weights = NULL, data = NULL, term.names = NULL,
 
     ## Coefficients
     b <- summary(m)$coef
-    b <- setNames(b[, 1], rownames(b))
+    if (isList(b)) b <- b[[1]]
+    if (is.matrix(b)) b <- setNames(b[, 1], rownames(b))
     xn <- names(b)
 
-    ## Intercept?
-    int <- isInt(xn)
-    int <- if (any(int)) {xn <- xn[!int]; TRUE} else FALSE
+    ## Intercept/no. parameters
+    int <- isInt(xn[1])
+    if (int) xn <- xn[-1]
+    k <- length(xn)
 
     ## Model weights
-    n <- nobs(m)
     w <- weights(m)
-    if (is.null(w)) w <- rep(1, n)
+    if (is.null(w)) w <- rep(1, nobs(m))
     s <- w > 0; w <- w[s]
 
+    ## Response
+    y <- getY(m)
+
     ## Centre/standardise x
-    k <- length(xn)
     if (k > 0) {
 
       ## Predictors
-      x <- model.matrix(m)[s, xn, drop = FALSE]
+      if (is.null(d)) d <- getData(m)
+      x <- model.matrix(m, data = d)[s, xn, drop = FALSE]
+      x <- data.frame(x, check.names = FALSE)
       obs <- rownames(x)
 
       ## Interactions?
@@ -1059,7 +1093,7 @@ stdCoeff <- function(mod, weights = NULL, data = NULL, term.names = NULL,
           xm <- colMeans(x)
 
           ## Adjust lower-order coefs
-          ## (ti = terms containing term i, ni = non-i components of ti)
+          ## (ti = terms containing term i; ni = non-i components of ti)
           b[xn] <- sapply(xn, function(i) {
             bi <- b[[i]]; XNi <- XN[[i]]
             ti <- xn[sapply(xn, function(j) all(XNi %in% XN[[j]])) & xn != i]
@@ -1074,21 +1108,20 @@ stdCoeff <- function(mod, weights = NULL, data = NULL, term.names = NULL,
           ## Centre predictors (for correct SD's/VIF's)
           if (std.x || unique.x) {
             x <- sapply(XN, function(i) {
-              xi <- x[, i, drop = FALSE]
-              xi <- sweep(xi, 2, xm[i])
+              xi <- sweep(x[i], 2, xm[i])
               apply(xi, 1, prod)
             })
+            x <- data.frame(x, check.names = FALSE)
           }
 
         }
 
-        ## Adjust intercept (weighted mean of predicted y)
+        ## Adjust intercept (set to weighted mean of predicted y)
         if (int) b[1] <- weighted.mean(predict(m, re.form = NA)[s], w)
-
       }
 
       ## Standardise by x
-      if (std.x) b[xn] <- b[xn] * apply(x, 2, sdW, w)
+      if (std.x) b[xn] <- b[xn] * sapply(x, sdW, w)
 
       ## Calculate unique effects of predictors (adjust for multicollinearity)
       if (unique.x && k > 1) {
@@ -1096,32 +1129,16 @@ stdCoeff <- function(mod, weights = NULL, data = NULL, term.names = NULL,
         ## Re-fit model with centred predictors
         ## (to calculate correct VIF's for interacting terms)
         m2 <- if (cen.x && inx) {
-
-          ## Add centred predictors to data
-          if (is.null(d)) d <- getData(m)
-          d <- c(list(x = x), as.list(d[obs, ]))
-
-          ## Add any offset(s)
-          o <- model.offset(model.frame(m)[s, ])
-          if (is.null(o)) o <- rep(0, n)
-          d <- c(list(o = o), d)
-
-          ## New model formula (rand. effs: https://bit.ly/2V1yDeu)
-          re <- if (isMerMod(m)) {
-            sapply(lme4::findbars(formula(m)), function(i) {
-              paste0("(", deparse(i), ")")
-            })
-          }
-          f <- reformulate(c("x", re), response = ".")
-
-          ## Update model
-          update(m, f, data = d, offset = o)
-
+          d <- d[obs, ]
+          xnc <- xn[xn %in% names(d)]
+          d[xnc] <- x[xnc]
+          d <- cbind(y, w, d)
+          update(m, y ~ ., weights = w, data = d)
         } else m
 
         ## Divide coefs by square root of VIF's
-        vif <- na.omit(VIF(m2))
-        b[xn] <- b[xn] / sqrt(vif)
+        vif <- VIF(m2, envir = environment())
+        b[xn] <- b[xn] / sqrt(vif)[xn]
 
       }
 
@@ -1129,29 +1146,30 @@ stdCoeff <- function(mod, weights = NULL, data = NULL, term.names = NULL,
 
     ## Centre/standardise y
     if (cen.y && int) {
-      ym <- weighted.mean(getY(m), w)
-      b[1] <- b[1] - family(m)$linkfun(ym)
+      ym <- weighted.mean(y, w)
+      if (isGlm(m)) {
+        f <- if (isBet(m)) m$link$mean else family(m)
+        ym <- f$linkfun(ym)
+      }
+      b[1] <- b[1] - ym
     }
     if (std.y) b <- b / sdW(getY(m, link = TRUE), w)
 
     ## Return standardised coefficients
-    sapply(xNam(m), function(i) unname(b[i]))
+    b <- sapply(xNam(m, d), function(i) unname(b[i]))
+    if (r.squared) c(b, R2(m, ...)) else b
 
   }
 
-  ## Add R-squared?
-  stdCoeff2 <- if (r.squared) {
-    function(m) c(stdCoeff(m), R2(m, d, ...))
-  } else stdCoeff
-
   ## Apply recursively
-  b <- rMapply(stdCoeff2, m, SIMPLIFY = FALSE)
+  b <- rMapply(stdCoeff, m, SIMPLIFY = FALSE)
 
   ## Output coefs or weighted average
   if (!is.null(w) && isList(b)) avgEst(b, w, bn)
   else {
     if (!is.null(bn)) {
-      rMapply(function(i) i[bn[bn %in% names(i)]], b, SIMPLIFY = FALSE)
+      f <- function(i) i[bn[bn %in% names(i)]]
+      rMapply(f, b, SIMPLIFY = FALSE)
     } else b
   }